TWEAK:  Attempt at making modules language-agnostic by using specIDs rather than class/spec names.  Potions probably won't work since I am hardcoding their names for the time being.  I can't test this myself.  :/
TWEAK:  Virmen's Bite Potion will only be recommended if the remaining duration of FET is >= 25 seconds, so it's not wasted.  Still deciding if I should just go back to SimC recommending it after the first minute (so that if you didn't pre-pot, you don't get annoyed by it until the second FET).