BUGFIX:  Cached textures shouldn't vanish on non-enUS clients.
BUGFIX:  Have to use localized potion names because the client doesn't know about the items on load.  (Ugh.)

WARNING:  You may experience an error loading this alpha on non-enUS clients.  If you have any errors that persist after re-selecting your active module, report it.  I was localizing the internal name of the "None" module, which meant that non-enUS clients have slightly different SavedVariables than enUS clients.  I've switched it back so that internal values are consistent, but what is displayed to the end-user is localized (if translation is available).