Hunter.lua - Added all necessary information for creating Hunter action lists. It models everything (to my knowledge) that a hunter needs to keep track of.
Events.lua - lastcast should not track auto shots. There is no reason to do so and interferes with tracking abilities like Steady Focus.

Hekili.toc - Added the hunter file to what wow needs to track.

State.lua - Added the fake buff pre_steady_shot to aid in predicting when you want to trigger the talent.
Also added the keyword cast_regen to track how much of a class resource will be generated when casting or waiting for gcd on an instant ability.